requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
click==8.1.8
distro==1.9.0
dnspython==2.7.0
fastapi==0.115.11
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
motor==3.7.0
openai==1.65.2
pydantic==2.10.6
pydantic_core==2.27.2
pymongo==4.11.2
python-dotenv==1.0.1
sniffio==1.3.1
starlette==0.46.0
tqdm==4.67.1
typing_extensions==4.12.2
uvicorn==0.34.0
